+++ sdk/communication/azure-communication-jobrouter/src/main/java/com/azure/communication/jobrouter/implementation/converters/WorkerAdapter.java
@@ -38,6 +38,7 @@ public class WorkerAdapter {
             .setTags(tags)
             .setAvailableForOffers(createWorkerOptions.isAvailableForOffers())
             .setChannels(createWorkerOptions.getChannels())
-            .setCapacity(createWorkerOptions.getCapacity());
+            .setCapacity(createWorkerOptions.getCapacity())
+            .setMaxConcurrentOffers(createWorkerOptions.getMaxConcurrentOffers());
     }
 }