+++ sdk/communication/azure-communication-jobrouter/src/test/java/com.azure.communication.jobrouter/RouterWorkerAsyncLiveTests.java
@@ -105,13 +105,15 @@ public class RouterWorkerAsyncLiveTests extends JobRouterTestBase {
             .setTags(tags)
             .setAvailableForOffers(false)
             .setChannels(channels)
-            .setQueues(queues);
+            .setQueues(queues)
+            .setMaxConcurrentOffers(1);
 
         // Action
         RouterWorker result = routerAsyncClient.createWorker(createWorkerOptions).block();
 
         // Verify
         assertEquals(workerId, result.getId());
+        assertEquals(1, result.getMaxConcurrentOffers());
         assertEquals(result.isAvailableForOffers(), false);
         assertEquals(labels.size() + 1, result.getLabels().size());
         assertEquals(tags.size(), result.getTags().size());